Separate tests to multiple files to run on different os
Golden tests must run on the os which generates the golden files.
The golden files was generated on windows
11 which is not supported by Github CI.
So I re-generated this files on macOS and updated the ci configuration.
